Move dynamic import to Safenet features
